% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{data}
\name{GSS_2010}
\alias{GSS_2010}
\title{GSS_2010}
\format{A data frame with 2044 rows and 5564 variables}
\source{
\url{http://www3.norc.org/GSS+Website/}
}
\description{
The 2010 cross-section of the General Social Survey.
}
\keyword{datasets}

