% Generated by roxygen2 (4.0.1): do not edit by hand
\name{Librarian}
\alias{Librarian}
\title{Load multiple packages}
\usage{
Librarian(x)
}
\arguments{
\item{x}{A character vector of package names.}
}
\description{
Load multiple packages
}
\details{
The \code{Librarian} function loads the packages named in \code{x}.
}
\examples{
my.packages <- c("car", "foreign")
Librarian(my.packages)
}
\author{
Jonah Gabry <jsg2201@columbia.edu>
}

