% Generated by roxygen2 (4.0.1): do not edit by hand
\name{logitOR}
\alias{logitOR}
\title{Table of odds ratios and confidence intervals for logit models}
\usage{
logitOR(fit, intercept = TRUE, level = 0.95)
}
\arguments{
\item{fit}{A logit model fit with \code{\link{glm}}.}

\item{intercept}{A logical value indicating whether the intercept should be included in the output.
Defaults to \code{TRUE}. If \code{fit} does not include an intercept the \code{intercept} argument
should be kept at the default value of \code{TRUE}.}

\item{level}{The desired confidence level to be passed to \code{\link{confint.glm}} (\pkg{MASS}).}
}
\value{
Odds ratios and confidence intervals for the parameters in \code{fit}.
}
\description{
Table of odds ratios and confidence intervals for logit models
}
\examples{
Y <- GSS_2010$realinc;  X <- with(GSS_2010, cbind(age, educ))
Y <- cut(Y, breaks=c(-Inf, median(Y, na.rm = T), Inf),
            labels=c("Low", "High"))
fit <- glm(Y ~ X, family = binomial)
logitOR(fit)
logitOR(fit, intercept = FALSE, level = 0.90)
}
\author{
Jonah Gabry <jsg2201@columbia.edu>
}

