% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{package}
\name{QMSS}
\alias{QMSS}
\alias{QMSS-package}
\title{Companion package for Greg Eirich's QMSS courses at Columbia University.}
\description{
Simple functions, data sets, example code, etc, for QMSS G4015 Quantitative
Analytic Techniques graduate course at Columbia University taught
by Greg Eirich.
}
\author{
Jonah Gabry <jsg2201@columbia.edu>
}

